+++ tests-data/struct-circular-ref-allowed/out.gen.d.ts
@@ -1,7 +1,7 @@
 import * as bare from "@bare-ts/lib"
 
 export interface Person {
-    readonly bestFriend: Person | undefined
+    readonly bestFriend: Person | null
     readonly secondBestFriend: 
         | { readonly tag: 0; readonly val: Person }
         | { readonly tag: 1; readonly val: undefined }